Remove ouroboros dev dependencies
Keep only self_cell
