+++ Cargo.toml
@@ -144,7 +144,6 @@ tempfile = "3.1.0"
 regex = "1.5.5"
 uuid = { version = "1.0", features = ["v4"] }
 unicase = "2.6.0"
-ouroboros = "0.18"
 self_cell = "1.1.0"
 # Use `bencher` over criterion because it builds much faster,
 # and we don't have many benchmarks
