+++ examples/owning_rows.rs
@@ -1,25 +1,23 @@
 extern crate rusqlite;
 
-use ouroboros::self_referencing;
 use rusqlite::{CachedStatement, Connection, Result, Rows};
+use self_cell::{self_cell, MutBorrow};
 
-#[self_referencing]
-struct OwningRows<'conn> {
-    stmt: CachedStatement<'conn>,
-    #[borrows(mut stmt)]
-    #[covariant]
-    rows: Rows<'this>,
-}
+type RowsRef<'a> = Rows<'a>;
+
+self_cell!(
+    struct OwningRows<'conn> {
+        owner: MutBorrow<CachedStatement<'conn>>,
+        #[covariant]
+        dependent: RowsRef,
+    }
+);
 
 fn main() -> Result<()> {
     let conn = Connection::open_in_memory()?;
     let stmt = conn.prepare_cached("SELECT 1")?;
-    let mut or = OwningRowsTryBuilder {
-        stmt,
-        rows_builder: |s| s.query([]),
-    }
-    .try_build()?;
-    or.with_rows_mut(|rows| -> Result<()> {
+    let mut or = OwningRows::try_new(MutBorrow::new(stmt), |s| s.borrow_mut().query([]))?;
+    or.with_dependent_mut(|_stmt, rows| -> Result<()> {
         while let Some(row) = rows.next()? {
             assert_eq!(Ok(1), row.get(0));
         }
